[NFC] Dump the built-in function name and arguments
Debug dumps of builtin functions were not showing what builtin functionality
was being used, or what the arguments were to the function. This adds a dump
function for the builtin function base class, that shows the name and the
arguments of the function.

This is only really useful for debugging, so no functional change.