Error on assignment of the result of divide and power to integer
The compiler would allow and compile something like

var a: integer;
...
  a := 2 / 5;
...

It would then store a double value into the integer variable, potentially
overwriting the next variable after. This is not allowed in Pascal, should
either use the div or some variant of conversion to integer.

Basic test cases added.

No change to valid Pascal code, this only affects code that was not
going to work correctly anyway.