Fix type compatibility for array types.
The type compatibility would accept any non-array type as compatible with
an array. This was due to a logic error in the CompatibleType function
where it would return the array type if the type wasn't an array.

This patch rearranges the return values in the CompatibleType function.

No change intended for valid Pascal programs, but some class of errors
are now caught that would previously not be detected.

A test case added to check that this is caught.